+++ src/main/java/org/jabref/model/entry/BibEntry.java
@@ -49,6 +49,7 @@ import org.jabref.model.util.MultiKeyMap;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.eventbus.EventBus;
 import com.tobiasdiez.easybind.EasyBind;
+import com.tobiasdiez.easybind.EasyBinding;
 import com.tobiasdiez.easybind.optional.OptionalBinding;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -1085,6 +1086,29 @@ public class BibEntry implements Cloneable {
 
         // Extract the path
         return FileFieldParser.parse(oldValue.get());
+    }
+
+    /**
+     * Gets a bound list of linked files.
+     *
+     * @param database maybenull
+     *                 The database of the bibtex entry.
+     * @return the list of linked files, is never null but can be empty.
+     * Changes to the underlying list will have no effect on the entry itself. Use {@link #addFile(LinkedFile)}
+     */
+    public EasyBinding<List<LinkedFile>> getFiles(BibDatabase database) {
+        OptionalBinding<String> result = getFieldBinding(StandardField.FILE);
+
+        // If there are no files on this entry we will fetch them from the referred entry.
+        if (result.isEmpty().get() && database != null) {
+            Optional<BibEntry> referred = database.getReferencedEntry(this);
+
+            if (referred.isPresent()) {
+                result = EasyBind.valueAt(referred.get().getFieldsObservable(), StandardField.FILE);
+            }
+        }
+
+        return result.mapOpt(FileFieldParser::parse).orElseOpt(Collections.emptyList());
     }
 
     public void setDate(Date date) {